pyproject.toml
[project]
name = "recipe-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohamed Khaled",email = "mohamed.khaled.200540@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "psycopg2 (>=2.9.11,<3.0.0)",
    "django (>=3.2.4,<3.3)",
    "djangorestframework (>=3.12.4,<3.13)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
